requirements.txt
Flask>=2.3.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
openpyxl>=3.1.2
Werkzeug>=2.3.7
pandas>=2.1.1
PyPDF2>=3.0.1
reportlab>=4.0.7
python-docx>=0.8.11
docxtpl>=0.16.7
SQLAlchemy>=2.0.21
Jinja2>=3.1.2
MarkupSafe>=2.1.3
click>=8.1.7
colorama>=0.4.6
itsdangerous>=2.1.2
python-dateutil>=2.8.2
pytz>=2023.3.post1
six>=1.16.0
tzdata>=2023.3
alembic>=1.12.0
Mako>=1.2.4
setuptools>=59.1.1
mysqlclient>=2.2.0
pillow>=10.0.0 
PyMySQL>=1.1.0 
cryptography>=45.0.0
zipfile36>=0.1.3
openpyxl-image-loader>=1.0.5

# 生产环境依赖
gunicorn>=21.2.0